Create ChocoChrome.txt
This is a Ducky Script that installs Google Chrome using Chocolatey package manager. The script starts by opening the Start menu on Windows 10 or 11, then inputs a command to install Chocolatey. After waiting for the installation to complete, the script inputs another command to install Google Chrome using Chocolatey. The script is named "ChocoChrome" and is authored by ooovenenoso.