Create ChocoTeamViewer.txt
This script installs TeamViewer using the Chocolatey package manager on a Windows computer. It starts by opening the Start menu and navigating to the Apps and Features option. Then it uses PowerShell commands to set the execution policy and download the Chocolatey installation script. After a delay, the script installs TeamViewer using the 'choco' command and confirms the installation by pressing the 'A' key.