Create ChocoAdobeReader.txt
The above code is a Ducky Script that automates the process of installing Adobe Reader using Chocolatey package manager on a Windows computer. The script begins by setting the execution policy to bypass and downloads the installation script from the Chocolatey website. After a delay of 15 seconds, the script proceeds to install Adobe Reader using the "choco install adobereader" command. Finally, the script sends the "A" key followed by the "Enter" key to accept any prompts and complete the installation. The author of this script is ooovenenoso.